Render iOS map

+++ google-maps/ios/Plugin/GoogleMapConfig.swift
@@ -3,31 +3,31 @@ import Capacitor
 
 // swiftlint:disable identifier_name
 public struct GoogleMapConfig: Codable {
-    let width: Int
-    let height: Int
-    let x: Int
-    let y: Int
+    let width: Double
+    let height: Double
+    let x: Double
+    let y: Double
     let center: LatLng
-    let zoom: Int
+    let zoom: Double
 
     init(fromJSObject: JSObject) throws {
-        guard let width = fromJSObject["width"] as? Int else {
+        guard let width = fromJSObject["width"] as? Double else {
             throw GoogleMapErrors.invalidArguments("GoogleMapConfig object is missing the required 'width' property")
         }
 
-        guard let height = fromJSObject["height"] as? Int else {
+        guard let height = fromJSObject["height"] as? Double else {
             throw GoogleMapErrors.invalidArguments("GoogleMapConfig object is missing the required 'height' property")
         }
 
-        guard let x = fromJSObject["x"] as? Int else {
+        guard let x = fromJSObject["x"] as? Double else {
             throw GoogleMapErrors.invalidArguments("GoogleMapConfig object is missing the required 'x' property")
         }
 
-        guard let y = fromJSObject["y"] as? Int else {
+        guard let y = fromJSObject["y"] as? Double else {
             throw GoogleMapErrors.invalidArguments("GoogleMapConfig object is missing the required 'y' property")
         }
 
-        guard let zoom = fromJSObject["zoom"] as? Int else {
+        guard let zoom = fromJSObject["zoom"] as? Double else {
             throw GoogleMapErrors.invalidArguments("GoogleMapConfig object is missing the required 'zoom' property")
         }
 
@@ -35,7 +35,7 @@ public struct GoogleMapConfig: Codable {
             throw GoogleMapErrors.invalidArguments("GoogleMapConfig object is missing the required 'center' property")
         }
 
-        guard let lat = latLngObj["lat"] as? Float, let lng = latLngObj["lng"] as? Float else {
+        guard let lat = latLngObj["lat"] as? Double, let lng = latLngObj["lng"] as? Double else {
             throw GoogleMapErrors.invalidArguments("LatLng object is missing the required 'lat' and/or 'lng' property")
         }
